Restructured the src/app folder
