+++ src/ws/WsEventListener.js
@@ -1,7 +1,7 @@
+import store from '../app/store';
 import modeNames from '../constants/modeNames';
 import { infoAlertDisplay } from '../features/alert/infoAlertSlice';
 import { boardPlayMove } from '../features/boardSlice';
-import store from '../store';
 import Wording from '../utils/Wording.js';
 import WsAction from '../ws/WsAction';
 import WsEvent from '../ws/WsEvent';
